requirements.txt
streamlit
pandas
numpy
plotly
networkx
pyyaml
statsmodels
seaborn
python-dotenv
matplotlib